fix: :bug: segfaulting on the test foreach and if its now theorically fixed thanks to alexandre
