+++ Makefile
@@ -36,7 +36,8 @@ SRC	=	./src/main.c								\
 		./src/builtins/fg.c							\
 		./src/builtins/alias.c						\
 		./src/builtins/unalias.c					\
-		./src/builtins/scripting/if.c 				\
+		./src/builtins/scripting/if/if.c 			\
+		./src/builtins/scripting/if/params.c 		\
 		./src/builtins/scripting/foreach/foreach.c 	\
 		./src/builtins/scripting/foreach/executor.c \
 		./src/builtins/scripting/repeat.c 			\
